+++ programs/opportunity_market/src/state.rs
@@ -227,7 +227,7 @@ pub struct OpportunityMarketOption {
     pub created_at: u64,
 
     pub total_staked: u64,
-    pub total_score: u64,
+    pub total_score: u128,
 
     pub reward_bp: Option<u16>,
 }
